Fix / Improve TextComponentInputStream and thus also CommandLineFrame
TextComponentInputStream used to not let the JVM terminate when the CommandLineFrame has been
closed via the frames close-button. This issue has been fixed by polling instead of continuously
blocking the thread while waiting for input.
Also we added the feature of deleting input, while we have to look out for the carets position,
which is now ensured with any input to be at the end of the TextComponent.