Minor  changes
Fixed the output of the help-command (and thus also the output of
Command.getDocumentation()) to consistently use platform-independent
line separator.

Adjusted and added tests for class Command (once because of the
previously mentioned fix, but also because of yet untested added features).

Refactored ICommandLine to be part of the
de.jeanpierrehotz.commandlinetools.command-package.