Refactor Command#getDosumentation() for more consistency and extensibility
Add some method-calls for creation of the first line of the documentation of a parameter
and for creation of the listing of valid values for enum parameters.
Also adds the feature of automatically listing all valid values for a enum parameter.
Furthermore updating the TODO.txt to be on the latest standing.