Fix a bug preventing the application to close under certain circumstances
As soon as the CommandLineFrame has been closed the Main-Application was going in a infinite loop
because the running-flag was still set to true but the streams (especially the InputStream) were closed
whereas the Scanner#nextLine()-method throws an exception.

This issue has been resolved by giving any ICommandLine a ICommandLineApplication which can then be
notified of the closing of the ICommandLine through the #onCLITermination()-method.