Add the possibility of giving parameters in form of a list
You can (and should) now assign Parameters an ordinal which determines in which
order values will be assigned to parameters. Parameters with a negative, or no
ordinal at all, will not be able to be given in list form. Any
parameters without default values *have* to have an ordinal assigned, while
it is optional for parameters with a default value.
Any parameter can be also explicitly given in any order, whereas they will simply
be skipped in the list.