Add ordinal to documentation and fix the build
The documentation has been updated to now also contain the ordinal of the single
parameters whereas also some tests needed to be adjusted.
Moreover I fixed the build which failed because when there was a parameter
given explicitly we tried to set the flag for the given parameter (in the list
representation) to have already been used. When this parameter doesn't exist
in the list representation though, we tried to set the flag at index -1 which
caused an undesired ArrayIndexOutOfBoundsException.