Refactor and document the new StringProcessing#tokenizeCommand(String)
As I rewrote the #tokenizeCommand(String)-method I didn't add any documentation
or comments to the code, which I havé  now done.
Also I have refactored the validation of the command with the validation of
the command name and of strings now being their own methods, which makes
the code more readable and also leads up to the already mentioned addition
of the ability to tokenize arrays.