Add possibility of executing shell commands
Now there is a distinction between a ShellCommand and CustomExecutableCommand, which
both implement the interface ExecutableCommand, although the way they are executed are
completely different.
The I/O for the ShellCommand is piped virtually through a file, and the execution
of it will only terminate as soon as the output of it has been completely read.