Fix bug preventing certain characters from being detected in the ui
Since characters written with ALT GR (like the backslash, which is used in
Windows as path separator) were ignored, but in a previous commit permitted.
This means that any character written with ALT GR shows up in the ui,
but is not saved / used by the InputStream.
This has been fixed by completely allowing any ALT GR-modified characters.